# swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  Jobs:
    properties:
      CreatedAt:
        description: Job creation time
        format: date-time
        type: string
      cat_id:
        description: Job Category ID
        type: string
        x-go-name: Cat_ID
      description:
        description: Description of the job
        type: string
        x-go-name: Description
      designation:
        description: job Designation or name
        type: string
        x-go-name: Designation
      id:
        description: ID of the job
        type: string
        x-go-name: ID
      location:
        description: Location of the job
        type: string
        x-go-name: Location
      org_id:
        description: ID of the organization posting the job
        type: string
        x-go-name: Org_id
      salary:
        description: Estimated salary of the job
        type: string
        x-go-name: Salary
    type: object
    x-go-package: _/home/sarmad/Go_Practice/PostJson/structures
  Memberships:
    properties:
      CreatedAt:
        description: created time
        format: date-time
        type: string
      org_id:
        description: Id of organization against which member is to be added
        type: string
        x-go-name: Org_ID
      pk:
        description: Membership ID
        type: string
        x-go-name: ID
      user_id:
        description: ID of the user to abe added as a member
        type: string
        x-go-name: U_ID
    type: object
    x-go-package: _/home/sarmad/Go_Practice/PostJson/structures
  OrgWrapper:
    properties:
      about:
        description: The email of the user
        type: string
        x-go-name: About
      name:
        description: The name of a user
        type: string
        x-go-name: Name
      pass:
        description: The password of the user
        type: string
        x-go-name: Website
      user_id:
        description: Organization's owner's ID
        type: string
        x-go-name: U_ID
    type: object
    x-go-package: _/home/sarmad/Go_Practice/PostJson/structures
  Organizations:
    properties:
      CreatedAt:
        description: Created Time
        format: date-time
        type: string
      about:
        description: About organization
        type: string
        x-go-name: About
      id:
        description: ID of organization
        type: string
        x-go-name: Org_ID
      name:
        description: Name of organization
        type: string
        x-go-name: Name
      user_id:
        description: Organization's owner's ID
        type: string
        x-go-name: U_ID
      website:
        description: Organization's website
        type: string
        x-go-name: Website
    type: object
    x-go-package: _/home/sarmad/Go_Practice/PostJson/structures
  Profile:
    properties:
      education:
        description: user's education
        type: string
        x-go-name: Education
      phone:
        description: user's Phone number
        type: string
        x-go-name: Phone
      user_id:
        description: ID of user of which, the profile is to be added
        type: string
        x-go-name: U_ID
    type: object
    x-go-name: ProfileWrapper
    x-go-package: _/home/sarmad/Go_Practice/PostJson/structures
  RequiredSkills:
    properties:
      job_id:
        description: Job ID against which skill is to be added.
        type: string
        x-go-name: Job_ID
      skill:
        description: skill name
        type: string
        x-go-name: Skill
    type: object
    x-go-package: _/home/sarmad/Go_Practice/PostJson/structures
  UserWrapper:
    properties:
      email:
        description: The email of the user
        type: string
        x-go-name: Email
      name:
        description: The name of a user
        type: string
        x-go-name: Name
      pass:
        description: The password of the user
        type: string
        x-go-name: Pass
    type: object
    x-go-package: _/home/sarmad/Go_Practice/PostJson/structures
  Users:
    properties:
      CreatedAt:
        description: User created time
        format: date-time
        type: string
      email:
        description: The email of the user
        type: string
        x-go-name: Email
      id:
        description: The uuid of a user
        type: string
        x-go-name: ID
      name:
        description: The name of a user
        type: string
        x-go-name: Name
      pass:
        description: The password of the user
        type: string
        x-go-name: Pass
    type: object
    x-go-package: _/home/sarmad/Go_Practice/PostJson/structures
  jobWrapper:
    properties:
      cat_id:
        description: Job Category ID
        type: string
        x-go-name: Cat_ID
      description:
        description: Description of the job
        type: string
        x-go-name: Description
      designation:
        description: job Designation or name
        type: string
        x-go-name: Designation
      location:
        description: Location of the job
        type: string
        x-go-name: Location
      org_id:
        type: string
        x-go-name: Org_id
      salary:
        description: Estimated salary of the job
        type: string
        x-go-name: Salary
    type: object
    x-go-name: JobWrapper
    x-go-package: _/home/sarmad/Go_Practice/PostJson/structures
  memberWrapper:
    properties:
      org_id:
        description: |-
          Organization ID
          in: body
        type: string
        x-go-name: Org_ID
      user_id:
        description: |-
          User ID
          in: body
        type: string
        x-go-name: U_ID
    type: object
    x-go-name: MemberWrapper
    x-go-package: _/home/sarmad/Go_Practice/PostJson/structures
host: localhost:5020
info:
  contact:
    email: srmdmjtba@gmail.com
    name: Sarmad Mujtaba
  title: Final Year Project - APIs.
  version: 0.0.1
paths:
  /category?category_id:
    get:
      description: 'This endpoint returns category specific jobs (1: Full-Time, 2:
        Part-Time, 3: Internship) if you pass that category ID as a query parameter'
      operationId: catParam
      parameters:
      - description: Category ID
        in: query
        name: category_id
        type: string
        x-go-name: Category_ID
      responses:
        "200":
          description: Jobs
          schema:
            $ref: '#/definitions/Jobs'
        "400":
          $ref: '#/responses/Error'
      summary: Lists all jobs of a category
      tags:
      - Jobs
  /jobs:
    post:
      description: You can add a Job through this endpoint by filling in the details
        of the job to be added and the organization which is posting the job.
      operationId: post-job
      parameters:
      - description: Add details of the user
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/jobWrapper'
        x-go-name: Body
      responses:
        "201":
          description: Jobs
          schema:
            $ref: '#/definitions/Jobs'
        "400":
          $ref: '#/responses/Error'
      summary: Add Job
      tags:
      - Jobs
  /jobs/designations?job_name:
    get:
      description: This endpoint returns designation specific jobs if you pass that
        designation's name as a query parameter
      operationId: nameParam
      parameters:
      - description: Organization ID
        in: query
        name: job_name
        type: string
        x-go-name: Job_Name
      responses:
        "200":
          description: Jobs
          schema:
            $ref: '#/definitions/Jobs'
        "400":
          $ref: '#/responses/Error'
        "404":
          $ref: '#/responses/Error'
      summary: Lists Jobs of given designation
      tags:
      - Jobs
  /jobs/skills:
    post:
      description: You can add multiple required skills for a job through this endpoint
        by filling in the details.
      operationId: post-requiredSkill
      parameters:
      - description: Add details of the user
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/RequiredSkills'
        x-go-name: Body
      responses:
        "201":
          description: RequiredSkills
          schema:
            $ref: '#/definitions/RequiredSkills'
        "400":
          $ref: '#/responses/Error'
      summary: Add Required Skill
      tags:
      - Jobs
  /jobs?id:
    delete:
      description: This endpoint deletes a Job if you pass its ID as a query parameter
      operationId: deleteJob
      parameters:
      - description: User ID
        in: query
        name: id
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/Error'
        "400":
          $ref: '#/responses/Error'
      summary: Delete Job
      tags:
      - Jobs
  /jobs?org_id:
    get:
      description: This endpoint returns all Jobs if no query parameter is passed.
        However, it returns organization specific jobs if you pass that organization's
        ID as a query parameter
      operationId: jobParam
      parameters:
      - description: Organization ID
        in: query
        name: org_id
        type: string
        x-go-name: ID
      responses:
        "200":
          description: Jobs
          schema:
            $ref: '#/definitions/Jobs'
        "400":
          $ref: '#/responses/Error'
        "404":
          $ref: '#/responses/Error'
      summary: Lists all / single Job
      tags:
      - Jobs
  /members:
    post:
      description: You can add a member of an organization through this endpoint by
        filling in the details.
      operationId: post-member
      parameters:
      - description: Add details of the user
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/memberWrapper'
        x-go-name: Body
      responses:
        "201":
          description: Memberships
          schema:
            $ref: '#/definitions/Memberships'
        "400":
          $ref: '#/responses/Error'
        "409":
          $ref: '#/responses/Error'
      summary: Add membership
      tags:
      - Memberships
  /members?org_id:
    get:
      description: This endpoint returns all members of an organizations based on
        organization's ID passed as query parameter.
      operationId: memberParam
      parameters:
      - description: Organization ID
        in: query
        name: org_id
        type: string
        x-go-name: Org_ID
      responses:
        "200":
          description: Memberships
          schema:
            $ref: '#/definitions/Memberships'
        "400":
          $ref: '#/responses/Error'
      summary: Lists Members
      tags:
      - Memberships
  /organizations:
    post:
      description: You can add an organization through this endpoint by filling in
        the details of the organization to be added.
      operationId: post-org
      parameters:
      - description: Add details of the Organization
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/OrgWrapper'
        x-go-name: Body
      responses:
        "201":
          description: Organizations
          schema:
            $ref: '#/definitions/Organizations'
        "400":
          $ref: '#/responses/Error'
        "404":
          $ref: '#/responses/Error'
      summary: Add organization
      tags:
      - Organization
  /organizations?id:
    delete:
      description: This endpoint deletes an organization if you pass its ID as a query
        parameter
      operationId: deleteOrgParam
      parameters:
      - description: Organization ID
        in: query
        name: id
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/Error'
        "400":
          $ref: '#/responses/Error'
      summary: Delete Organization
      tags:
      - Organization
    get:
      description: This endpoint returns all organizations if no query parameter is
        passed. However, it returns single organization if you pass its ID as a query
        parameter
      operationId: orgParam
      parameters:
      - description: Organization ID
        in: query
        name: id
        type: string
        x-go-name: ID
      responses:
        "200":
          description: Organizations
          schema:
            $ref: '#/definitions/Organizations'
        "400":
          $ref: '#/responses/Error'
        "404":
          $ref: '#/responses/Error'
      summary: Lists all / single organizations
      tags:
      - Organization
  /profile:
    put:
      description: You can add a user profile through this endpoint by filling in
        the details of the user to be added.
      operationId: post-profile
      parameters:
      - description: Add details of the user
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/Profile'
        x-go-name: Body
      responses:
        "200":
          description: Profile
          schema:
            $ref: '#/definitions/Profile'
        "201":
          description: Users
          schema:
            $ref: '#/definitions/Users'
        "400":
          $ref: '#/responses/Error'
        "409":
          $ref: '#/responses/Error'
      summary: Add Profile
      tags:
      - Profile
  /profile?user_id:
    get:
      description: This endpoint returns a user's Profile if you pass its ID as a
        query parameter
      operationId: userProfile
      parameters:
      - description: User ID
        in: query
        name: user_id
        type: string
        x-go-name: User_ID
      responses:
        "200":
          description: Users
          schema:
            $ref: '#/definitions/Users'
        "400":
          $ref: '#/responses/Error'
      summary: Get User's Profile
      tags:
      - Profile
  /users:
    post:
      description: You can add a user through this endpoint by filling in the details
        of the user to be added.
      operationId: post-user
      parameters:
      - description: Add details of the user
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/UserWrapper'
        x-go-name: Body
      responses:
        "201":
          description: Users
          schema:
            $ref: '#/definitions/Users'
        "400":
          $ref: '#/responses/Error'
        "409":
          $ref: '#/responses/Error'
      summary: Add User
      tags:
      - User
  /users?id:
    delete:
      description: This endpoint deletes a user if you pass its ID as a query parameter
      operationId: deleteParam
      parameters:
      - description: User ID
        in: query
        name: id
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/Error'
        "400":
          $ref: '#/responses/Error'
      summary: Delete User
      tags:
      - User
    get:
      description: This endpoint returns all users if no query parameter is passed.
        However, it returns single user if you pass its ID as a query parameter
      operationId: userParam
      parameters:
      - description: User ID
        in: query
        name: id
        type: string
        x-go-name: ID
      responses:
        "200":
          description: Users
          schema:
            $ref: '#/definitions/Users'
        "400":
          $ref: '#/responses/Error'
        "404":
          $ref: '#/responses/Error'
      summary: Lists all / single users
      tags:
      - User
produces:
- application/json
responses:
  Error:
    description: ""
schemes:
- http
swagger: "2.0"